pyproject.toml
[project]
name = "moex_iss_sdk"
version = "0.0.1"
description = "Typed MOEX ISS client shared by MCP servers."
authors = [{ name = "team-moex-agent" }]
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.6",
    "fastmcp>=2.13.3",
    "python-dotenv>=1.0",
    "prometheus-client>=0.20",
]

[project.optional-dependencies]
dev = ["pytest>=7.4"]

[tool.setuptools.packages.find]
where = ["."]
